Allow options longer than 1025 chars
Calico [1] with OpenStack uses dnsmasq with a very long
--bridge-interface option:
--bridge-interface=<context-if-name>,<alias-if-name>,<alias-if-name>,...,
where each occurrence of ",<alias-if-name>" occupies 15 characters, and
there can in principle be as many <alias-if-name>s as you can have VMs
on a single OpenStack compute host.

Currently the option is limited to 1025 chars overall, which only allows
for 67 <alias-if-name>s, which is not necessarily enough.  This patch
removes the limit by reallocating the arg buffer when necessary.

[1] https://docs.projectcalico.org/v2.6/getting-started/openstack/